setting all data types to tf.float64

+++ python/exp01_gemm.py
@@ -12,8 +12,8 @@ def gemm_implicit_noup(A, B, C):
 
 def exp01_gemm(b, n):
 
-    A = tf.random.normal([n, n])
-    B = tf.random.normal([n, n])
-    C = tf.random.normal([n, n])
+    A = tf.random.normal([n, n], dtype=tf.float64)
+    B = tf.random.normal([n, n], dtype=tf.float64)
+    C = tf.random.normal([n, n], dtype=tf.float64)
 
     res = b.benchmark('gemm_implicit_noup', gemm_implicit_noup, A, B, C)
